Add ability to comment and delete all comments
