+++ portfolio/src/main/webapp/index.html
@@ -6,9 +6,10 @@
 
   <head>
     <meta charset="UTF-8">
-    <title>My Portfolio</title>
+    <title>Payton Dennis Portfolio</title>
     <link rel="stylesheet" href="style.css">
     <script src="script.js"></script>
+    <script src="todos.js"></script>
   </head>
 
   <!-- NOTE: Whenever body loads, we will update the STATE of our function. React would be really helpful right now :) -->
@@ -62,6 +63,7 @@
         <!-- Used to append todos to element! -->
         <p>Todos:</p>
         <ul id="todos"></ul>
+        <button type="button" onclick="deleteTodos()">Delete All Tasks!</button>
       </div>
 
       <br>
