+++ portfolio/src/main/webapp/script.js
@@ -45,55 +45,3 @@ async function addFoodToDOM() {
 
     document.getElementById('fooditems').innerText = data + "!"; 
 }
-
-
-/** Function that fetches the state of the todo array and appends them to unordered list todos. */
-async function updateTodos() {
-    // Refresh Comments
-    refreshTodos();
-
-    // Grab the # of Comments
-    const todoQuantity = document.getElementById("quantity").value;
-
-    // authenticate quantity given
-    if (authenticateQuantityGiven(todoQuantity)) {
-
-        // Fetch that Quantity
-        const response = await fetch(`/todo?amount=${todoQuantity}`);
-        const data = await response.json();
-        console.log(data);
-
-        createTodos(data);
-
-    } else {
-        console.log("Error! Please enter another value!");
-    }
-}
-
-/** Create Todos */
-function createTodos(data) {
-    const TodoNode = document.getElementById('todos'); 
-    for (todo of data) {
-        // Create li element
-        const liElement = document.createElement('li');
-        liElement.innerText = todo;
-
-        TodoNode.appendChild(liElement);
-    }
-}
-
-/** Refresh Comments */ 
-function refreshTodos() {
-    const TodoNode = document.getElementById('todos');
-    while (TodoNode.firstChild) {
-        TodoNode.removeChild(TodoNode.firstChild);
-    }
-}
-
-function authenticateQuantityGiven(todoQuantity) {
-    // set max here
-    if (todoQuantity > 0 && todoQuantity <= 5) {
-        return true;
-    } else
-        return false;
-}
